Created modules and switched filtering with already existing items. Asynced Delete action.
