+++ constants.js
@@ -1,0 +1,2 @@
+export const URL = "http://localhost:3000/items";
+export const categoryAll = ["All items"];
