+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Green</title>
-    <script src="app.js" defer></script>
+    <script type="module" src="app.js" defer></script>
     <link rel="stylesheet" href="style.css" />
     <link
       rel="stylesheet"
